@@ -7,14 +7,6 @@
     ],
     "homepage": "https://github.com/graze/skeleton-project",
     "license": "MIT",
-    "config": {
-        "platform": {
-            "ext-simplexml": "1.0",
-            "ext-dom": "1.0",
-            "ext-tokenizer": "1.0",
-            "ext-xmlwriter": "1.0"
-        }
-    },
     "authors": [
         {
             "name": "__author-name__",
@@ -29,10 +21,10 @@
         }
     ],
     "require": {
-        "php": "^5.5 | ^7.0"
+        "php": "^5.6 | ^7.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.2 | ^5.2",
+        "phpunit/phpunit": "^5.7.21 | ^6 | ^7",
         "squizlabs/php_codesniffer": "^3.0",
         "graze/standards": "^2.0"
     },
@@ -46,6 +38,11 @@
             "Graze\\Skeleton\\Test\\": "tests/src",
             "Graze\\Skeleton\\Test\\Unit\\": "tests/unit",
             "Graze\\Skeleton\\Test\\Integration\\": "tests/integration"
+        }
+    },
+    "config": {
+        "platform": {
+            "php": "7.2"
         }
     }
 }
