@@ -9,11 +9,13 @@
  * file that was distributed with this source code.
  *
  * @license https://github.com/graze/skeleton-project/blob/master/LICENSE.md
- * @link https://github.com/graze/skeleton-project
+ * @link    https://github.com/graze/skeleton-project
  */
 
 namespace Graze\Skeleton\Test;
 
-class ExampleTestCase extends \PHPUnit_Framework_TestCase
+use PHPUnit\Framework\TestCase as PHPUnitTestCase;
+
+class TestCase extends PHPUnitTestCase
 {
 }
